Move video uploads and datastore to common
